Create Object Tools: Show preview already on hover
Now the object to be placed appears on hover instead of on clicking,
which provides feedback about what's going to happen when you click.

This was a rather tricky change since it affects some complex existing
bit of code. I expect there to be some bugs to be ironed out.

Closes #537