Optimized deletion of many objects
Deleting many selected objects, either directly or as part of undoing
the addition of those objects, was very slow. This was due to each
object getting removed individually and the large amount of code getting
executed in response to the related change of selected objects.

I've optimized this by supporting the removal of many objects at once in
the undo command, so that it can remove those objects from the selection
in one operation.

Closes #1972