Use Qt 5.9 LTS instead of 5.10 for releases
Qt 5.9 is an LTS release that is receiving more bugfix releases. Since
we're not really using any Qt 5.10 or 5.11 features there is little
reason not to use this stable version of Qt. At the same time, there are
regressions and issues affecting Tiled functionality in newer versions.

This changes the Linux AppImage to Qt 5.9.6 and the Windows release to
Qt 5.9.5 (5.9.6 is not available yet on AppVeyor).

Closes #1928