Allow interaction with objects overlapping other maps
This patch is a rather tricky workaround to allow interacting with
objects and related handles that are overlapping another map. Clicking
this other map would switch to that map, but by accepting hover events
in the graphics items that sit on top we can avoid that and allow the
move and resize interactions to be done there as well.

Since this affects the mouse handling in a bunch of places, we need to
keep an eye open for possible regressions.

This patch also fixes the problem that you could not interact in areas
of the map below which another map was present. The current map now
steals the hover events from the map below, so that the map below knows
when to ignore the mouse press event.

Closes #1983