Integrated news from website in the status bar
Current there is no option to turn this off. Not sure if this is
important, but it could be added to the preferences.

I didn't find a nice way to display the date of the posts, but I think
it doesn't really matter.

I'm not sure whether it is preferable to add the itch.io devlog or the
news feed from the Tiled website. The itch.io devlog currently is more
active since it includes posts about snapshot updates. On the other
hand, I have more control over the posts on the Tiled website and the
way they are presented. I went for showing the posts on the Tiled
website for now, and will consider posting snapshot updates there too,
as well as improving the site a bit.

Closes #1898