Properties view now applies changes to all selected layers
When having multiple layers selected, changing properties like the
visibility or locked status now apply to all selected layers instead of
just the last selected one.

The same goes for custom properties, which are also aggregated based on
all selected layers, as was already implemented for objects.

Part of issue #282