Made "Offset Layers" tool work on all selected layers
The tool makes sure that when a parent layer is moved, its children are
not moved as well, to avoid applying the offset twice.

Also made some other improvements, to reduce the chance of accidental
use:

* Only drag on left mouse button, not on right mouse button.
* Show a "SizeAll" / hand cursor when pressing the mouse button.
* While dragging, allow right mouse button or Escape to abort the drag.

Also fixed a problem with updating a layer's opacity when "Highlight
current layer" is active and multiple layers are selected.

Part of issue #282