Fixed disappearing of tile type on export/import
When exporting or importing a tileset, or when enabling any export
options that affect the tileset, the tileset is cloned. In the process
of cloning however, the tile types were not getting copied over.

See also change 8494cd6d4a02f4799c81e473a82048b5bd6b8bbe. Maybe there is
a way to avoid making such errors so easily?

Closes #2023