Fixed tool shortcuts when using Spanish translation
Two tools were using the shortcut "G" (Wang Brush and Eraser) and two
tools were using the shortcut "R" (Rectangular Select and the Bucket
Fill Tool). When this happens, neither shortcut will actually work.

Resolved the conflicts by reverting to the same shortcuts used in the
English language, which is anyway a good idea since those are the
shortcuts mentioned in the manual.

See https://discourse.mapeditor.org/t/keyboard-shotcuts-not-work/3506