Travis CI: Use BUTLER_API_KEY for authentication
Less of a hassle than decrypting a file and silences a warning that
recommends using the environment variable.

The "travis" command-line tool reformatted the entire .travis.yml. I
opted to let it have its indentation preference and hope it didn't break
anything.

Also changed from using the bleeding-edge version of butler to the
stable version, in an attempt to avoid a crash on macOS.