Expose MapDocument.map as a property
This makes the map, along with its properties, available to JavaScript.

Direct access to the map currently allows making changes without
notifying the UI or involving the undo system. This easily leads to
errors, so further work is required on the API for making changes. The
access to the raw data should probably be read-only.

Example snippet that changes the tile width, which explicitly notifies
the UI but still ignores the undo system:

    var mapDocument = tiled.documentManager.currentDocument
    mapDocument.map.tileWidth = 16
    mapDocument.mapChanged()