Scripting: Further work on the exposed API
Added tiled.activeAsset property and tiled.openAssets to get a list of
EditableAsset instances. Added EditableLayer class and used it to expose
the layers from EditableMap and allow changing some basic properties.

This enables a quick and easy API for changing properties:

  > tiled.activeAsset.layerCount
  45
  > tiled.activeAsset.layerAt(2).name
  Terrain
  > tiled.activeAsset.layerAt(2).visible = false
  false

All while being integrated with the undo system and automatically
updating the UI based on the changes.