Scripting: Added actions to modify the selection
Also fixed the enum-based map properties, though they are still
sub-optimal since they currently just work as integers.

Regarding the selection, there is still no way to process it. Also, I'd
rather have exposed the selection as a value-type but functions
modifying such a value-type are not working as expected when the value
is exposed as a property. Maybe I'll change it to selectedArea() and
setSelectedArea(), when no better solution is found.