Scripting: Exposed tile layer and tileset properties
Tile layer size and individual cells can now be queried, though there is
no way yet to change the tile layer.

All simple tileset properties are now exposed and some of them can be
changed directly (name, tileOffset and backgroundColor). No access to
the tiles is provided yet.

Also it is now possible to remove layers using map.removeLayerAt(index).
However, this also introduces a problem if you try to change a layer
that was removed. This will probably just need to be checked against.