Scripting: Added a value-type for working with regions
Exposed EditableTileLayer.region() method for getting the non-empty
region of a tile layer. The returned value can be modified and can be
used to change the selected area of the map.

A lot of overloads were necessary because "Qt.rect" actually creates a
QRectF, and without supporting that specific class the script engine was
not picking the right overload.

Currently it's not possible to create a region object directly. This
should be possible by wrapping its static QMetaObject, but I ran into
crashes when trying to do "new Region()".