Scripting: Resolved several potential crashes in layer handling
* When a layer is removed from the map, any EditableLayer instance for
  that layer is invalidated.

* Trying to change something about an invalid EditableLayer writes an
  error to the console (from Qt 5.12, throws a script error and includes
  a stack trace in the message when applicable). This is mostly necessary
  because there is no undo stack to apply the change through.

* EditableLayer instances can now be garbage-collected. When this happens,
  their cached pointer is removed so that the instance is re-created
  next time it is requested.

A remaining problem is to use an EditableLayer instances for a layer
that was not only removed but _deleted_. Then reading out any properties
will still crash. Either the invalidation should clear the layer pointer
as well, or the EditableLayer instance should keep the layer from being
deleted, so that it could be used to add the layer again.