Scripting: Added Map.insertLayerAt and modifying of detached layers
This method currently allows adding a removed layer back to a map.

Also, it is now possible to modify layers that are not currently part of
a map. In this case, change are applied immediately without involving
the undo system. The layer is cloned when it is removed from the map, to
avoid any changes being visible when undoing the removal of the layer.

This change also resolves the possible crash when accessing properties
of a deleted layer. This is because before a layer gets deleted, it
would first need to get removed from a map. From that point on, the
properties are obtained from the clone instead.

Now we have a small issue, in that EditableTileLayer instances managed
by the scripting engine are keeping a TileLayer instance around, which
in turn keeps its tilesets around. This currently triggers an assert in
the TilesetManager destructor. This could be fixed by explicitly
destroying the ScriptManager before the TilesetManager.