Removed the "Internal" namespace
Never seen noticed any benefit from it, so good riddance!

I kept it only for two classes in libtiled, which use the "Internal"
namespace to separate a private class.

This change will affect the context of many translatable strings, but
this mostly just causes them to be marked as "unfinished" since they are
still matched by heuristic.