Added 'json1' plugin that exports to the old JSON format
For those that need to use the old JSON format (usually for compatiblity
reasons), there is now a plugin called 'json1' that can be activated
along with or instead of the regular 'json' plugin.

The JSON format was simplified in Tiled 1.2, but not everybody is in a
position to update the software they are using to adjust to the new
format.

Closes #2058