Scripting: Allow creating new stuff from scripts
This change enables the following constructors from script:

    new MapObject([name])
    new ObjectGroup([name])
    new TileLayer([name])   # Not very useful yet
    new TileMap()           # Not useful yet
    new Tileset([name])     # Not useful yet

Newly created map objects can be added to new or existing object groups,
and newly created object groups can be added to an existing map.

Tilesets can't yet be added to the map, and it's also still not possible
to actually make a non-empty tileset.

Tile layers aren't very useful yet because you still can't modify them.

The map class is called 'TileMap' because in JavaScript, the global name
'Map' is taken.