tBIN plugin: Fixed hang when locating missing tileset image
Was caused by trying to cut the tileset into 0x0 tiles, of which there
are infinitely many.

Now the plugin uses Tileset::findOrCreateTile to create the tiles for
loading the tile properties, which does not affect the tile size
(Tileset::addTiles does, because it's meant to be used by image
collection tilesets).

Closes #2068