Scripting: Added initial support for scripted tools
Scripted tools can change whatever is possible to change from scripts
based on input events.

Some important features are still missing however, like the ability to
show a preview of a change to tile layers or of the object the tool
plans to insert. Currently it's also not possible to set an icon or a
shortcut for the tool, and registered tools are only available in the
Map Editor.