tBIN plugin: Fixed possible crash when images can't be found
A crash would occur when a tileset image failed to load and the map
contained animated tiles from the affected tileset.

This change also makes a few small optimizations (avoiding copying and
length calculations in some places) and improves the rendering of
missing tiles in case those tiles had properties or an animation
associated with them.

Also fixed saving of tileset properties.

Closes #2106