Made keyboard shortcuts configurable (#2111)
This currently covers most actions present in the menus, but still 
excludes those that are only available through a shortcut.

Shortcuts can be set, removed and reset back to their default. Changed 
shortcuts are stored in the user preferences. The shortcuts can also be 
exported to and imported from an XML file.

Closes #215