JSON plugin: Fixed writing of position for objects without ID
Normally all objects have an ID, except for object templates (where we
do not want to write a position), so this will rarely have been an
issue. Where it went wrong, is when you load an old map where objects as
part of tile collision shapes did not have IDs assigned yet, because if
you don't edit the tile collision shapes those objects will not get IDs
assigned. This in turn caused their position to not be saved.

Fixed by adjusting the check for whether an object is an object
template.

Forum-topic: https://discourse.mapeditor.org/t/tmx-and-json-object-data-doesnt-seem-to-match/3899