docs: Clarify the meaning of the "staggered" orientation (#2118)
There are three uses of staggered tiles: Staggered Isometric, Staggered Hexagonal, and Staggered Orthogonal (which is normal squares, but each tile has 6 neighbors instead of 9). So the name `staggered` could refer to either the Isometric or the Orthogonal meanings (hexagonal is clearly documented as using staggered axes in addition to the `staggered` orientation).