Fixed the image cache to check file modification time
This fixes problems for image files that are not watched for changes
(like the images used by image layers), or when the watching for changes
is disabled by the user. In that case, a restart of Tiled or explicit
reloading of the tilesets using the hidden Ctrl+T shortcut was needed
before images would be reloaded. Manually changing the image source and
then back, or just reloading the map was not enough since the image
would simply be taken from the cache.

The new behavior is that images will be reloaded whenever they are found
in the cache, but with an out of date "last modified" timestamp.

Closes #2081