Tile Animation Editor: Update preview scale immediately
Instead of waiting until we have to update the current frame, the scale
of the preview is now changed immediately along with the zoom factor of
the tileset view.

Also fixed the scale to be applied when the preview is reset as well.

Finally, the horizontal size hint of the preview label was set to ignore
so that high zoom factors don't break the layout.

Issue #2050