Worlds: Fixed handling of negative multiplierX/Y
It was invalidating the pattern since the mapWidth/mapHeight properties
defaulted to values of multiplierX/multiplierY, without taking the
absolute value, and negative map sizes are invalid.

So, a workaround for those affected would be to explicitly set the
mapWidth/mapHeight properties to the appropriate positive value.

Tiled-forum: https://discourse.mapeditor.org/t/negative-value-in-world-offset-value/3963