Scripting: Provide read/write access to tile collision shapes
To enable this, the last relevant change signal,
MapDocument::objectGroupChanged, was changed to the
ObjectGroupChangeEvent, allowing an EditableObjectGroup and the
EditableMapObject to be changed so that they only need to refer to an
EditableAsset rather than an EditableMap.

This means EditableTile::objectGroup can now return an
EditableObjectGroup, and the Tile Collision Editor can use the change
events to make sure its copy of the tile's object group is re-created
when necessary.

What doesn't work, is the "selected" property of map objects that are
part of the tile's object group, since object selection still requires a
MapDocument, and there is no way to access to the dummy document used by
the Tile Collision Editor, which anyway uses a copy of the object group.
To solve this, ideally there would not be a copy, but to avoid that
would require way too many other changes at this point.

Issue #949