Fixed exporting to a file name with multiple dots
For example, you could not export a map as "map-0.0.1.json" because it
would try to find a map format that used a "0.1.json" file extension.

As a workaround for this issue, the format can be explicitly selected by
using the file type filter in the file dialog. Then Tiled does not
search for a matching format.

Closes #2149