Added search filter to the Objects view
Thanks to @thabetx for working on the initial implementation of this
in pull request #1566. This patch takes some ideas from there, but it's
rewritten from scratch.

The biggest difference is that after changing the implementation of the
ReversingProxyModel to be based on QSortFilterProxyModel, this patch can
now apply the filter directly instead of needing to use another proxy
model. This makes the patch a lot smaller.

Closes #1467
Closes #1566