Simplified the implementation of ReversingProxyModel
Instead of implementing an entire custom proxy that attempts to reverse
the children, the new implementation just relies on
QSortFilterProxyModel and does a reverse sort on the row IDs.

This might be a bit slower theoretically, since sorting was not actually
required, but it avoids a lot of tricky code and removes bugs in that
code that affected future patches.

It's also nice to be able to apply a filter to it without needing to
chain together multiple proxy models.