Added Issues view for displaying warnings and errors
The plan is for this new view to replace existing error reporting
mechanisms in some places and to provide a place where errors reported
from scripts can be collected.

Each error only shows up once, with duplicates being counted (currently
not displayed, but is planned to be).

The script function tiled.error() now displays the error both in the
Console and the Issues view, and a new function tiled.warn() was added
to create warnings.