Extended the use of the new Issues view
The following issues now show in the Issues view as well:

* Errors and warnings reported by Python scripts
* Errors output by custom commands
* Warnings previously written to standard-error in the tBIN plugin
* New warning in tBIN plugin when animation frame durations differ

This change also significantly simplifies the way the logging is
connected to the UI, by making the LoggingInterface a singleton instead
of adding instances of it to the PluginManager.

The AutoMapper uses new system to make its warnings and errors show up
in both the Console and the Issues view.

Issues with callbacks reported from a script now get counted twice, so
probably more refactoring is needed.