Report issues for missing tilesets and templates
Double-clicking the issues opens a file chooser for locating the files
elsewhere, and replaces the tileset / template for the map in which the
issue happened (this may be a little problematic, since multiple maps
may need their references corrected).

I wanted to replace the broken links widget with this, but it does not
cover all the functionality of that widget, and I'm not sure how to make
it do that. In particular the ability to select a directory to search
for all missing files is hard to give a new place.