Issues view: Show warnings for custom file path properties
A warning is now added if the custom file path property points to a file
that does not exist.

Double-clicking the warning will generally select the relevant object,
bring the Properties view into focus and select the custom property,
though this has not been implemented for all cases (see added todos).

Closes #2080