Save the last export target and format in the map/tileset file
Saving this information in the file allows the export to be repeated
without prompt even after restarting Tiled. This is supported for both
TMX and JSON formats.

At the same time, the export format is now remembered by its short name
rather than a smart pointer, which makes sure that when the right format
can still be picked after loading/unloading a plugin or changing an
export script.

The fallback to TMX or TSX formats has also been removed, since it would
only cause problems to silently export in these formats when a
previously used format can't be found.

The export target is stored with a relative path, so that the setting
may be generally useful also when the file is shared.

Issue #1610