Scripting: Added initial access to tile stamp and tile edit preview
The current tile stamp can now be accessed via the
MapEditor.currentBrush property. If the tile stamp has multiple
variations, this property returns the first one. To change the tile
stamp, a map can be assigned to this property. Support for stamps with
multiple variations may be added later.

A scripted tool can now set a tile edit preview via the
ScriptedTool.preview property. Soon it will also be possible to
conveniently merge the preview to actually apply the change.

Issue #949