pytiled: A Tiled Python extension library
This is just an initial start, but the idea is to have a Python
extension library that provides access to some basic Tiled
functionality, which can be used without starting the editor.

It may also be able to replace the API provided by the Tiled Python
plugin, though right now I'm not sure yet how that could be set up.