Scripting: Resolved garbage collection issues
If a QObject that was passed to the JavaScript engine is still alive,
that does not mean that it can always still be turned into a non-null
QJSValue. Hence, before we return a cached QObject instance, we need to
check for this.

Also explicitly set QQmlEngine::JavaScriptOwnership on these objects,
because when they are returned from a Q_PROPERTY the ownership would
otherwise default to CppOwnership and the object would never get deleted
(which resulted in tileset instances still being around on shutdown).