pytiled: Separated the Python extension module from the Python plugin
This way the distribution may be a little more complex, but it would
open up the possibility of using the Python scripting API in stand-alone
Python scripts.

It's still a bit of a mess though and I'm not sure if this setup would
work well in practice.

This removes the pybind11 bits as well and keeps the PyBindGen related
code for now, since it would be a lot of work to translate all the API
to pybind11. It also avoids another build-time dependency.