Scripting: Added support for custom binary map formats
The 'toString' member for scripted map formats has now been renamed to
'write' and may return either a string or an ArrayBuffer object.

When an error happens during export a backtrace is now reported in the
Console and Issues views.

Also, no longer fall back to "Export As" when triggering "Export" runs
into an error. Turns out opening the file dialog after the error dialog
is just annoying.

Issue #949