Fixed undo for terrain brush on previously empty area
Since change 6ca522e1e7b96802a9484d7a467ced2918a0d577, the source tile
layer used for TileLayer::setCells can have 0 width and height, but when
a non-empty region is given alongside that means we want to erase that
region.

Now there exists two overloads for TileLayer::setCells, one that works
as before based on the size of the given layer, and one that always sets
all cells in the given area. There did not appear to be a use-case for
doing the intersection.