Display Qt warnings and errors in Console and Issues views
This change intercepts any debug messages, warnings and error messages
sent through the Qt library in order to display them in the Console and
Issues views.

It may be confusing because some of these messages may not be anything
the user can do something about, but this seems like the only way to get
script errors that happen while calling functions connected to signals
to be visible. We'll see if there's any messages that should rather be
filtered out.