Fixed visible area changing when reloading map
This seemed to be mainly due to the current view state not being saved
yet when the reloaded map was opened. Since closing the old map first
causes an unwanted additional switch of document, the state is now saved
explicitly before opening the new version.

Also, the new map is not immediately inserted at the right index, rather
moving the tab to the right place afterwards.

Finally, a rounding error was avoided when restoring the center of the
view, making sure it stays absolutely stable upon reload.

Closes #2213