Make shortcuts no longer translatable
I currently know of no actual usage of this feature to change a shortcut
in a particular language. The only thing this seems to have done is to
cause confusion and extra work for translators who will translate the
names of the keys, which causes the shortcuts to break since Qt no
longer knows which key is meant.

If there are any keys that need adjustment based on language or keyboard
layout, it's probably best to handle them explicitly somewhere.