Scripting: Changed the way file formats write to files
Rather than handling the return value as either a string or an
ArrayBuffer to be written to a file, the script can use the TextFile or
BinaryFile API to write to the file by itself. This also enables export
formats that write several files, so the 'outputFiles' function is now
available for custom map formats.

When the 'write' function returns a non-empty string it is now shown as
error message.

Also made TextFile and BinaryFile use QSaveFile when writing, which adds
the need to call commit() when done but avoids loss of data in case of
disk errors or code errors.

Also fixed BinaryFile.OpenMode enum to be actually exposed to the
script.

Issue #949