Tiled Quick: Don't compile/install by default
When building with qmake, you now need to pass CONFIG+=tiled_quick to
get the tiledquickplugin and tiledquick to build.

When building with Qbs, tiledquickplugin and tiledquick will only be
built when mentioned specifically using the --products/-p command-line
argument. Generally tiledquickplugin does not need to be mentioned
because it will build as a dependency of tiledquick.

Also fixed the installation path on UNIX to be $${PREFIX}/bin instead of
installing into /opt, when using qmake.

Finally raised the minimum version of Qt for tiledquickplugin to 5.6,
since it's the minimum version supported by libtiled.

Closes #2673