Layers view: Fixed insertion position when using drag-n-drop
Due to the change in how the layer list is reversed, the row in
LayerModel::dropMimeData needed to be adjusted differently.

Also optimized it a little bit, since I noticed when dragging a set of
layers, each layer was reparented three times (to the same parent, so it
wasn't really noticable). It happened once for each column in the model.