Simplified Project class, moved scanning to ProjectModel
The model seems a better fit for the scanning, especially taking into
account future asynchronous scanning.

Also, the model now keeps track of the name filter and updates it as
appropriate when file formats get added or removed. With a timer, to
prevent excessive scanning on startup and after reloading of scripts.