Introduced sessions
The session is currently implicit, being either the default session or a
project-based session saved alongside a project.

It stores the list of open files, the active file, recent files and the
last state for previously opened files. This makes it easier to switch
between different projects, which swaps all these things along with it.

Also improved the start location for the open file dialog, which now
starts at the active file, then the first recent file and finally the
user home directory (instead of starting at the current working
directory, usually at the binary).

Issue #1665