Support tmx files as readable images
This allows for users to create metatiles, by using a map file as the
source image of a tileset. The same code also allows ANY image to be a
rendered version of a map file, which may allow for interesting use cases.

Maps that can be used this way have some restrictions:
- They must be Orthogonal.
- Object layers are allowed, but they can only contain tile objects.

See discussion in the original issue #1421 for the full pros/cons of this
method.