docs: Removed repeated listing of built-in formats
There are really only two built-in formats, which are the 'tmx' and
'tsx' formats. All the others are added by plugins which can be disabled
by the user.

Regardless, I don't see much value in documenting this list of formats
and then having to maintain those lists (which will surely be
forgotten).