Fixed object creation tools and a number of corner cases
* Fixed alignment handling for shape objects for isometric renderer
* Fixed alignment handling for object creation tools
* Added alignment handling to text objects

Objects that are in the process of being created are not yet part of the
map, so the renderer has to pass its map into MapObject::alignment to be
able to get a correct alignment that takes into account the
objectAlignment property of the map or the map's orientation when it is
not set.

Tile and text insertion tools want to center the object at the mouse
position and need to take into account the alignment of the object once
it is added to the map to place it correctly.

Rectangle and ellipse insertion tools need to adjust the bounds they set
on the created object to take into account the alignment as well.