Fixed displayed layer data format in Properties
When the map was using CSV layer data format and Zstandard compression
was not supported, it would display as XML layer data format in the
Properties.

Also more clearly marked the XML format as deprecated. For performance
and file size reasons it should not be used when it can be avoided.