New objects tree view partly works
It crashes when trying to edit a pre-existing value, probably because a
QModelIndex is being constructed with the default no-arg constructor.

There is a visibility checkmark for all layers and objects that should be
removed, and a position column that should be removed. The resize rules
also need to be updated so that name and type get priority. ID should be
the first column, if possible, as well.

Layers can be selected, which shouldn't happen.

The filter bar is broken.