Introduced Tileset.alignment property
This property specifies the alignment to use for the tiles in the tileset,
overriding the default behavior implemented in the map renderer.

The map object alignment was extended to support all alignment values.
If set, it overrides the alignment specified in the tileset.

The intention is for this property to be applied for tiles on tile
layers as well.