Fixed context menus activating first item on release
Added some margin in menu panels to work around the issue where the first
action in the menu would sometimes appear below the mouse and get
immediately activated on release, even when not moving the mouse.

This issue was not reported for Tiled 1.2 so it seems to be new in
Tiled 1.3. Potentially it's affected by some change between Qt 5.9 and
Qt 5.12. It appears to only affect Windows.

Closes #2693