GitHub Actions: Made macOS workflow upload the Tiled.app
This Tiled.app is untested so it may have some problems.

Also, for some reason it ships with a bunch of Qt frameworks that are
not normally included and should not be necessary, like QtQuick,
QtVirtualKeyboard and QtQmlModels. This might be related to building
against Qt 5.14 rather than 5.12.

Finally, the Tiled version did not end up in the Info.plist for some
reason.