Scripting: Fixed crashes when modifying certain new objects
Because the Document class now owns the undo stack (for a moment it has
been the EditableAsset class), we need to ensure the asset has a
document associated with it before trying to make changes using undo
commands.

Also added a few missing read-only checks to setProperty, setProperties
and removeProperty.

Closes #2747