Fixed crash on shutdown
This was a regression introduced by f511b75ea7a799372f1bc9e45b9f2e4cb92.
The DocumentManager destructor was referencing widget pointers that got
deleted already now that it was being called later.

I've introduced a custom TabBar class (which also replaces the
WheelEnabledTabBar) to avoid the need to register an event filter for
making tabs closeable using middle-click.

The remaining mWidget pointer is now wrapped in a QPointer, which will
automatically become null when the widget is already deleted as part of
the UI hierarchy.