Fixed loading of compression level
When there is no compression level set in a map file, it should default
to -1 (default compression) and not 0 (minimal compression).

In addition, the compression level is no longer written out for TMX maps
that use the default level.

Also added the compressionlevel attribute to the documentation.
Although, I don't like to have it on the <map> element, it should be
moved into the <editorsettings> element.

Closes #2753