Scripting: Prevent crash in Asset.macro/undo/redo/isModified
Assets don't always have an associated undo stack, but these script
functions were not checking for it (because this wasn't the case when
they were written).

Now, calling Asset.undo() or Asset.redo() raises a script error if
called on an asset without undo stack. Asset.isModified() always returns
false in this case. Asset.macro() just doesn't start a macro.

Also documented the Asset.undo() and Asset.redo() functions, as well as
adjusting the documentation to mention that modifications to assets are
only done via the undo system if they are loaded in the editor.