Fixed problems with saving the default session
If you were not using predefined object types or had not saved any tile
stamps, then the data directory used to store the default session did
not exist yet and saving would fail.

Make sure to always create the data directory now that we're storing the
default session there.

Thanks to sverx for reporting this issue on Discord.