Improved map positioning when toggling 'Clear View'
Now it keeps the center of the view in the center when the map view is
resized as a result of toggling the 'Clear View' action, reducing the
amount the map shifts around.

I decided not to set the resize anchor to AnchorViewCenter permanently
because this causes a lot of potentially distracting movement when
resizing the window or the dock widgets.