Scripting: Removed assertion that could trigger from script API
The Q_ASSERT(!tileSize.isEmpty()) in Tileset::setTileSize might have
made sense in the past, but now scripts can set up the tile width and
height independently and we can't assert a valid size at either point.

A width or height of 0 is only invalid when we try to cut an image into
tiles anyway, so interrupting a script that is just modifying this
property is likely not helpful either.