Scripting: Fixed loading of images from extensions folder
It is now possible to set the image of a tileset to for example
"ext:tileset.png", if a Tiled extension ships with such a tileset.

While a tileset's image could already be set to this value, loading the
image failed because of the way the URL was converted to a string.  Now
it checks for the "ext" prefix.