Remember the preferred format used for saving
There are several formats that Tiled can read and write, which show up
in the Save As dialog. For new files, the default format was always TMX
for maps and TSX for tilesets. Now it will be the last used format.

This change is especially meant to make life easier for people
preferring the JSON format.