Scripting: Fixed map.tilesets when 'Embed tilesets' is enabled
When the option to embed tilesets on export is enabled, then tilesets
which are not already embedded are replaced with an embedded copy before
the 'write' function of a custom map format is called.

These copies did not have associated TilesetDocument instances and as
such did not have an EditableTileset instance that could be returned for
this property.

Fixed by creating temporary EditableTileset instances on-demand.

Closes #2767