Session: Moved the automapping while drawing option to session
I was planning to move more options to the session, but I'm undecided
whether it is an improvement so it's just this option for now.

Since session options may change when the session changes, a callback
mechanism was introduced so that UI elements can update when either the
option or the session is changed.