Project: Added project-specific extensions folder
By default a folder called "extensions" in the same folder as the
.tiled-project file will be added to the extensions path. This folder
can be changed in the new Project Properties dialog.

The script engine will reset when switching projects also causes a
change in the extensions path, as well as when changing the extensions
path of the current project.

Issue #1665