Project: Automatically refresh the file system view
Watch the folders for changes and automatically trigger a (slightly
delayed) re-scan of changed folders.

Ideally, only actually changed folders would need to be re-scanned, but
that could be an optimization for later. Currently a top-level folder is
entirely re-scanned when any file or directory is added or removed
within in.

Also fixed that selection and vertical scroll position is not lost when
refreshing the view, by restoring it after each refresh.

Issue #1665