Fixed initialization and restoring of map view
In some scenarios, most prominently on Windows when Tiled starts up
maximized, the map view did not have its final size at the time the map
is scaled and positioned. This caused maps opened for the first time
through the command-line or the explorer to be scaled down too much and
also caused maps that were open previously to not be in the place they
were when Tiled was closed.

Worked around this issue by delaying the initialization of scale and
position until the first paintEvent.

Closes #2779