Added status bar button to toggle the Console view
Since the Console and Issues views share the same space by default, I
made them soft-mutually-exclusive. When one of them is made visible
while tabbed together with the other one, the other one is automatically
hidden. This seems more convenient than having the two buttons on the
status bar be simple toggles and then having the tabs in addition to
toggle between the views.

I had to disable grouped dragging for the top-level main window,
otherwise it was impossible to separate the Console and Issues views
from each other.